[src] Add GruNonlinearityComponent; moving around some sources in nnet3 to avoid very large files.
